requirements.txt
psycopg[binary]
python-dotenv
Flask
